split tests setup into db and client files
